requirements/base.txt
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-cors-headers==4.4.0  # https://github.com/adamchainz/django-cors-headers
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-extensions==3.2.3  # https://github.com/django-extensions/django-extensions
django-redis==5.4.0  # https://github.com/jazzband/django-redis
djangorestframework==3.15.2  # https://github.com/encode/django-rest-framework
djangorestframework-simplejwt==5.3.1
# DRF-spectacular for api documentation
drf-spectacular==0.27.2  # https://github.com/tfranzel/drf-spectacular
drf-yasg==1.21.7
ipython==8.26.0
kafka-python-ng==2.2.3
redis==5.0.7  # https://github.com/redis/redis-py
requests==2.32.3
Werkzeug==3.1.3
